+++ unique_orchestrator/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2025-10-30
+- Add option to customize the display of tool progress statuses.
+- Make follow-questions postprocessor run last to make sure the follow up questions are displayed last.
+
 ## [1.6.1] - 2025-10-28
 - Removing unused experimental config `full_sources_serialize_dump` in `history_manager`